# docs/openapi.yml
openapi: 3.0.0
servers:
  - description: Production
    url: http://localhost:9090
info:
  description: Codewind API
  version: 0.1.0
  title: Codewind API
  contact:
    name: API support
    email: example@ibm.com
  license:
    name: Microclimate Developer License
    url: 'https://github.ibm.com/dev-ex/codewind/blob/master/LICENSE/LICENSE.md'

# tags:
  # - name: users
  #   description: Operations available to regular users
  # - name: admin
  #   description: Operations available only to authenticated admins
paths:
...
  /api/v1/projects/{id}/open:
    put:
      summary: Open project
      parameters:
        - name: id
          in: path
          schema:
            $ref: '#/components/schemas/ProjectID'
          required: true
          description: id of project
      responses:
        200:
          description: Project successfully opened
        404:
          description: Project with given projectID not found
...
  /api/v1/registry:
    get:
      summary: Get the status of the Deployment Registry in the Workspace Settings File
      responses:
        200:
          description: Successfully completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: The response status code
                    example: 200
                  deploymentRegistry:
                    type: boolean
                    description: true if deployment registry was set
                    example: true  
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: The response status code
                    example: 400
                  deploymentRegistry:
                    type: boolean
                    description: false if deployment registry was not set
                    example: false
    post:
      summary: Test the specified Deployment Registry 
      description: The API also emits a socket event `deploymentRegistryStatus` with data `{deploymentRegistryTest ":" true, msg ":" message related to the socket for pass/failure}`
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - deploymentRegistry
              properties:
                deploymentRegistry:
                  type: string
      responses:
        200:
          description: Successfully completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: The response status code
                    example: 200
                  deploymentRegistryTest:
                    type: boolean
                    description: true if deployment registry is valid
                    example: true
                  msg:
                    type: string
                    description: Message related to the request
                    example: Success message 
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: The response status code
                    example: 400
                  deploymentRegistryTest:
                    type: boolean
                    description: false because it is a bad request
                    example: false
                  msg:
                    type: string
                    description: Message related to the request
                    example: Reason for failure
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: The response status code
                    example: 500
                  deploymentRegistryTest:
                    type: boolean
                    description: false if deployment registry is invalid
                    example: false
                  msg:
                    type: string
                    description: Message related to the request
                    example: Reason for failure
...
components:
  schemas:
...
    ProjectID:
      type: string
      format: uuid
      example: d290f1ee-6c54-4b01-90e6-d701748f0851
      pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}'
...